Provide means of submitting jobs to a drained queue
During an upgrade, the job queue needs to be drained in order to avoid
new jobs coming to the cluster.  Nevertheless, the upgrade process
needs to carry out some maintenance, like redistributing the new
configuration, therefore, this patch provides a means of submitting
jobs to a drained queue.

Of course, once the more fine-grained job queue control will be implemented,
this functionality can be removed.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Jose Lopes <jabolopes@google.com>