Add instance disk fields
The instance disk fields have been added. As these are the first
fields to be duplicated, some helper functions are present to support
easier specification of fields which have to be duplicated for every
entry in an array of configuration objects, and numbered accordingly.
A newtype of Maybe was introduced to fix serialization issues, as well
as a minor utility function.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Jose A. Lopes <jabolopes@google.com>