Optimize dd parameters for instances moves
Benchmarks have shown that increasing the buffer size and disabling
synchronous output increased the throughput by about 10MiB/s, so make
those parameters the default.

Note that not using sync for output doesn't increase the risk of data
loss, as the data will remain on the source node anyway if an error
occurs during the move. This change, and increasing the block size to 1
MiB, didn't increase the memory consumption of dd during the move
operation significantly.

Signed-off-by: Thomas Thrainer <thomasth@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>