Add the first instance query live fields
This patch adds live fields which can be implemented with the existing
RPC calls. Some fields require additional call types and will be added
later. Most of them can be added at this point.
The status field is especially interesting, as it depends not on the
information returned, but whether it was returned, and where from.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Jose A. Lopes <jabolopes@google.com>