Create check-news Makefile target
Checking the correctness of the NEWS file syntax is an important part of the
check process, but up to now it was only possible as part of a bigger set of
tests. This commit creates a Makefile target to run that independently.

The developer notes are updated to document this new target.

Signed-off-by: Michele Tartara <mtartara@google.com>
Reviewed-by: Jose A. Lopes <jabolopes@google.com>