Watcher cleans user shutdown domains
Update watcher to destroy shutdown domains when it detects that a
domain is in the state 'USER_down'.  This requires adding logic
similar to that of restart instance, that is tracking the number of
cleanup attempts in the watcher's state and give up if a given number
of retries has been reached.

Signed-off-by: Jose A. Lopes <jabolopes@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>