Update QA to test custom SSH ports
This patch adds the `ssh-port` option. If set to a non-standard port,
the QA script sets up the default node group with this port, and before
running tests it adds `iptable` rules to all nodes so that the nodes
see each other's SSH servers as running on this port.
Their SSH configuration is _not_ changed and other machines see the
nodes on 22 as before.

The `iptable` rules are reset on each QA run, trying to preserve any
existing rules (not created by the script) that might be present.

Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>