In the list of involved nodes, drop "no secondary"
When grouping moves into jobs, a new job set is started, if the new
move involves a node also touched by a previous move. When computing
the list of involved nodes, the new primary and secondary nodes of the
instance are included; if an instance, however, has only one node, the
"no secondary" index -1 should be left out to avoid spurious conflicts
between unrelated moves of single-homed instances.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>