Implement 'QueryInstances' call in Haskell luxi server
While the command line uses the generic 'Query' call,
rapi calls 'QueryInstances'. 'QueryInstances' so far
was not fully implemented in the Haskell implementation
of the luxi server. This was discovered when trying to
switch from python to haskell queries. This patch
fixes the Haskell luxi server implementation wrt to that.

Signed-off-by: Helga Velroyen <helgav@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>