Destroy Xen domain only if necessary
Check if a given Xen domain still exists after a 'xm shutdown' before
destroying it.  This check is necessary in order to cope with Xen
domains created with other version of Ganeti, given that an instance
created in or prior to Ganeti 2.10 will be cleaned up automatically
and thus calling 'xm destroy' is unnecessary and will lead to an
error, whereas instances from the more recent versions of Ganeti do
require 'xm destroy' as it is part of the instance shutdown detection
feature.

Signed-off-by: Jose A. Lopes <jabolopes@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>