Remove duplicate storage units in node query
This is a little performance tweak for the node queries.
So far, the query code mapped the disk templates to storage
units. It could happen that two disk templates were mapped
to the same storage unit and therefore the storage space
information was calculated twice for the same unit. This
patch removes this unecessary double calculation by
filtering for duplicate storage units before invoking the
the RPC call.

Signed-off-by: Helga Velroyen <helgav@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>