Use hypervisor / storage information only when requested
So far, the node queries ignored the list of fields and
just requested all available information from the backend.
That means, for example if only hypervisor information is
requested, still the storage space calculation is
triggered. This patch makes the node queries aware of
their fields and only calls the RPC with the minimum
necessary parameters.

Signed-off-by: Helga Velroyen <helgav@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>