Add various single instance operations
To further expand the number of RAPI methods in the workload, the
single instance operations are added in this patch. An instance is
created, deleted, shutdown, restarted, reinstalled, renamed, and
has its disks activated and deactivated and grown.

As this increases the complexity of the code, similarly themed RAPI
operations have been extracted into separate functions.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>