Activate QA for rapi queries via luxi
This patch enables QA testing for rapi queries for the
newly transformed queries from python to haskell
(groups, instances, nodes, export, and networks). So far,
the QA did not distinguish between resources that cannot be
queried with PUT requests, hence an additional distinction
between luxi-able and luxi-and-put-able resources was
necessary.

Signed-off-by: Helga Velroyen <helgav@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>