Design: Improvements of Node Security
This is a design doc addressing issue 377. Objective is
to reduce the number of nodes that are able to establish
ssh and RPC connections to other nodes. Limiting this
set of nodes to the master candidates is desired to
decrease the risk of a compromised node compromising the
entire cluster.

Signed-off-by: Helga Velroyen <helgav@google.com>
Reviewed-by: Guido Trotter <ultrotter@google.com>