design: scheduling will be simple in the first iteration
In order to progress faster towards jobs as processes, the decision on
the final scheduling algorithm will be postponed. For the time being,
the simple one will be used that starts jobs until a fixed maximal
number of running jobs is reached. This postponing also has the
advantage that the final decision can be based on experience with the
characteristics of jobs as processes under realistic load.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>