Add per-resource query workload
The query requests are done to receive data about a certain resource
type. With tests for all the resources barring networks in place, the
query workloads can be added at the point where the existence of enough
resources in the system can be confirmed, making the results of the
queries meaningful.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>