Don't assume we win the archive race
The job scheduler in luxid regularly watches for changes
of the job files to determine progress of jobs. As these
files are updated atomically, reading them will always
succeed---until they're archived. While luxid is quite
fast observing queue changes, it still is possible that
a job gets archived before luxid noting that it finished.
Make luxid aware of this possibility.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>