gnt-cluster verify: demote orphan volume error to warning
Ganeti checks for orphan volume by making sure that it knows about all
volumes on disk; any additional orphan volume, even if created by the
administrator, causes a failure in gnt-cluster verify. Given that
hail and hbal already account and warn for missing space, this is
unnecessarily strict.

This commit demotes this error to a warning.

This commit also updates the part of QA about orphan and reserved
logical volumes to match, adapting the QA infrastructure slightly.

Fixes Issue 314.

Signed-off-by: Santi Raffa <rsanti@google.com>
Signed-off-by: Michele Tartara <mtartara@google.com>
Reviewed-by: Michele Tartara <mtartara@google.com>