Allow the skipping of checks for inter-cluster move test
The inter-cluster instance move test is very interesting for the RAPI
compatibility tests, as it uses many RAPI requests that are otherwise
hard to exercise. It uses no command-line functionality apart from
the standard QA instance checks.

As the move-instance tool does its own checks of all the prerequisites
and desired results, not using the QA checks is acceptable for the RAPI
workload. This patch allows the checks to be skipped via an optional
argument.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>