Make the move-instance tool more fault tolerant
The move-instance tool raises an exception when used with a cluster
running an earlier version of Ganeti. As the tool is meant to perform
inter-cluster moves, this situation could be encountered outside the
RAPI compatibility test.

To handle it better, this patch makes the tool fail more gracefully,
outputting a more informative error message.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>