Add warning about the RecreateInstanceDisks invocation
A test relying on RAPI alone cannot exercise the RecreateInstanceDisks
functionality properly - simply because it cannot damage an instance
to the point where its disks would be missing and in need of
recreation. The test is left standing simply to test if the operation
is still exposed via the interface.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>